Created Service layer for Questions and Answer. Created some endpoints in Question and Answer controllers. Faced Internal server issues (quizId null), needs fixing

+++ src/main/java/com/sda/repositories/QuizRepository.java
@@ -3,6 +3,6 @@ package com.sda.repositories;
 import com.sda.model.quizzes.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizRepository extends JpaRepository<Quiz,Integer> {
+public interface QuizRepository extends JpaRepository<Quiz,Long> {
     Quiz findByQuizTitle(String quizTitle);
 }
